In run time (not in plan time), I changed the implementation of Run() such that we won't use Expand() method each time, but only on the first time of each batch. After the first time, we allocate a Data array of size equals to the width of the data from the first iteration.

**PPROF** results, before and after, can be seen in the PR's [Task](https://app.asana.com/0/573768021211412/1123190355912217)

A possible improvement to the PR is to implement it in **plan time**