@BoraBak 2019-05-16
**Benchmark**
----------------------------------------------------------------------------------
@aviaoh I think my benchmark isn't good enough. Can you please have a look...

**Code:**
```go
func BenchmarkProjectImpl(b *testing.B) {
	runner := ep.Project(&question{}, &upper{}, &lower{}, &question{}, &count{})
	data := ep.NewDataset(strs([]string{"sUndAy", "mOnDay", "TueSdAY", "wednESday", "tHuRSday", "FRIday"}))
	for n := 0; n < b.N; n++ {
		data, err := eptest.Run(runner, data)
		require.NoError(b, err)
		require.NotNil(b, data)
	}
}
```

- **Before refactoring:**
    - `BenchmarkProjectImpl-4             30000             41202 ns/op            4795 B/op        120 allocs/op`
- **After refactoring - option#1:** (before Avia's code changing suggestion)
   - `BenchmarkProjectImpl-4             30000             43420 ns/op            4875 B/op        121 allocs/op`
- **After refactoring - option#2:**
   - `BenchmarkProjectImpl-4             30000             36737 ns/op            4794 B/op        118 allocs/op`