@aviaoh 2019-05-16
sure:)

> 	runner := ep.Project(&question{}, &upper{}, &lower{}, &question{}, &count{})

should result with an error of mismatched rows. instead, define project without count, where all runners return the same amount of rows.

>		data, err := eptest.Run(runner, data)

you are testing project with single batch. i.e. optimized part is never tested.
instead, define 
```go
	data := ep.NewDataset(strs([]string{"sUndAy", "mOnDay", "TueSdAY", "wednESday", "tHuRSday", "FRIday"}))
	batches := make([]ep.Dataset, 1000)
	for i:=0; i<1000; i++ {
		batches[i] = data
	}
	b.Reset
	for n := 0; n < b.N; n++ {
		data, err := eptest.Run(runner, batches...)
		...
	}