@BoraBak 2019-05-16
**Benchmark - round 2**
----------------------------------------------------------------------------------

**Code:**
```go
func BenchmarkProjectImpl(b *testing.B) {
	runner := ep.Project(&question{}, &upper{}, &question{}, &lower{}, &question{})
	data := ep.NewDataset(strs([]string{"sUndAy", "mOnDay", "TueSdAY", "wednESday", "tHuRSday", "FRIday"}))
	batches := make([]ep.Dataset, 1000)
	for i:=0; i<1000; i++ {
		batches[i] = data
	}
	b.ResetTimer()
	for n := 0; n < b.N; n++ {
		data, err := eptest.Run(runner, batches...)
		require.NoError(b, err)
		require.NotNil(b, data)
	}
}
```

- **Before refactoring:**
    - `BenchmarkProjectImpl-4               100          14648209 ns/op         2465517 B/op      70107 allocs/op`
- **After refactoring - option#1:** (before Avia's code changing suggestion)
   - `BenchmarkProjectImpl-4               100          13145823 ns/op         2225983 B/op      64114 allocs/op`
- **After refactoring - option#2:**
   - `BenchmarkProjectImpl-4               100          13066646 ns/op         2225604 B/op      64110 allocs/op`

@aviaoh