@aviaoh 2019-05-16
ok cool :)
can you please commit the new version to this pr? 
also, we should check data length by same code as dataset.expand.
i.e.:
```go
func hasDifferentLength(set1, set2 Dataset) bool {
	// when expanding with variadicNulls - don't force same length
	thisLen := set.Len()
	otherLen := other.Len()
	isAnyVariadicNulls := thisLen < 0 || otherLen < 0
	return !isAnyVariadicNulls && thisLen != otherLen
}
```

i was hoping to see larger improvement here, so lets try profile this benchmark with most updated code.