@BoraBak 2019-05-19
**Benchmark - round 3**
----------------------------------------------------------------------------------

**Code:**
```go
func BenchmarkProjectImpl(b *testing.B) {
	b.Run("2runners", func(b *testing.B) {
		runner := ep.Project(&upper{}, &question{})
		data := ep.NewDataset(strs([]string{"sUndAy", "mOnDay", "TueSdAY", "wednESday", "tHuRSday", "FRIday"}))
		batches := make([]ep.Dataset, 1000)
		for i := 0; i < 1000; i++ {
			batches[i] = data
		}
		b.ResetTimer()
		for n := 0; n < b.N; n++ {
			data, err := eptest.Run(runner, batches...)
			require.NoError(b, err)
			require.NotNil(b, data)
		}
	})
	b.Run("5runners", func(b *testing.B) {
		runner := ep.Project(&question{}, &upper{}, &question{}, &lower{}, &question{})
		data := ep.NewDataset(strs([]string{"sUndAy", "mOnDay", "TueSdAY", "wednESday", "tHuRSday", "FRIday"}))
		batches := make([]ep.Dataset, 1000)
		for i:=0; i<1000; i++ {
			batches[i] = data
		}
		b.ResetTimer()
		for n := 0; n < b.N; n++ {
			data, err := eptest.Run(runner, batches...)
			require.NoError(b, err)
			require.NotNil(b, data)
		}
	})
	b.Run("100runners", func(b *testing.B) {
		runners := make([]ep.Runner, 100)
		for i := 0; i < 96; i++ {
			runners[i] = &question{}
			runners[i+1] = &upper{}
			runners[i+2] = &question{}
			runners[i+3] = &lower{}
			runners[i+4] = &question{}
		}
		runner := ep.Project(runners...)
		data := ep.NewDataset(strs([]string{"sUndAy", "mOnDay", "TueSdAY", "wednESday", "tHuRSday", "FRIday"}))
		batches := make([]ep.Dataset, 1000)
		for i:=0; i<1000; i++ {
			batches[i] = data
		}
		b.ResetTimer()
		for n := 0; n < b.N; n++ {
			data, err := eptest.Run(runner, batches...)
			require.NoError(b, err)
			require.NotNil(b, data)
		}
	})
}

```

- **Before refactoring:**
    - `BenchmarkProjectImpl/2runners-4                      200           6,894,873 ns/op          913,136 B/op      28,052 allocs/op`
    - `BenchmarkProjectImpl/5runners-4                      100          14,757,872 ns/op         2,465,085 B/op      70,106 allocs/op`
    - `BenchmarkProjectImpl/100runners-4                     10         199,460,147 ns/op        130,009,814 B/op   1,211,826 allocs/op`
- **After refactoring:**
   - `BenchmarkProjectImpl/2runners-4                      200           6,548,417 ns/op          913,253 B/op      28,054 allocs/op`
   - `BenchmarkProjectImpl/5runners-4                      100          13,178,912 ns/op         2,225,435 B/op      64,108 allocs/op`
   - `BenchmarkProjectImpl/100runners-4                     20         100,191,696 ns/op        43,818,863 B/op    1,015,821 allocs/op`