@BoraBak 2019-05-21
> 👌
> please change also `composeProject` + benchmark it

New [PR](https://github.com/panoplyio/ep/pull/166) is opened