@aviaoh 2019-05-21
great, so waiting only for comment `please use it also in dataset.expand`  above :)